+++ boilerplate/web_modules/layouts/Page/index.js
@@ -3,10 +3,6 @@ import { PropTypes } from "react"
 import Helmet from "react-helmet"
 import invariant from "invariant"
 
-// function pageDescription(text) {
-//   return text
-// }
-
 export default class Page extends Component {
 
   static propTypes = {
@@ -44,11 +40,12 @@ export default class Page extends Component {
       { property: "og:type", content: "article" },
       { property: "og:title", content: metaTitle },
       { property: "og:url", content: __url },
-      // { property: "og:description", content: pageDescription(body) },
+      { property: "og:description", content: head.description },
       { name: "twitter:card", content: "summary" },
       { name: "twitter:title", content: metaTitle },
       { name: "twitter:creator", content: `@${ pkg.twitter }` },
-      // { name: "twitter:description", content: pageDescription(body) },
+      { name: "twitter:description", content: head.description },
+      { name: "description", content: head.description },
     ]
 
     return (