+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Added: Generate meta description and bundle it to collection
+  See ``boilerplate/web_modules/layouts/Page/index.js`` for an example.
+  ([#79](https://github.com/MoOx/statinamic/issues/79))
 - Changed: build and client scripts now require an exports value
   that must contains layouts, metadata and routes.
   This values should be strings.